kernel pca gets its own module
